+++ src/mjolnirtools/__init__.py
@@ -1,3 +1,8 @@
 """Convenience tools for users of the Mjolnir HPC cluster."""
 
-__version__ = "1.1.2"
+from importlib.metadata import version, PackageNotFoundError
+
+try:
+    __version__ = version("mjolnirtools")
+except PackageNotFoundError:
+    __version__ = "unknown"